refactor(audio-listeners): split initAudioListeners into per-concern setup modules
Each listener/subscription concern moves into its own module under
store/audioListenerSetup/; initAudioListeners just composes them in the
original setup / teardown order:

- audioEngineListeners  audio:* + analysis:* Tauri listeners
- initialAudioSync      one-shot startup sync to the Rust engine
- authSyncListener      auth-store + analysis-storage subscriptions
- mprisSync             MPRIS / OS media-controls sync
- radioMprisMetadata    radio ICY StreamTitle -> MPRIS
- discordPresence       Discord Rich Presence sync

Verbatim code-move — listener registration, handler bodies, local state
and the exact cleanup order are all preserved. The MainApp call site and
the three playerStore.*.test.ts suites import initAudioListeners
unchanged.
